requirements.txt
# S&OP 智能决策中枢 v3.0 / LangChain 学习项目
python-dotenv>=1.0.0

langchain>=0.3.0
langchain-classic>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0

faiss-cpu>=1.8.0
rank-bm25>=0.2.2

pandas>=2.0.0
tabulate>=0.9.0
httpx>=0.27.0
pydantic>=2.0.0
